// src/app/generated/commander/model/movedMediaData.ts
/**
 * plexhelp-commander API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MediaDescriptionData } from './mediaDescriptionData';
import { MediaFileType } from './mediaFileType';


export interface MovedMediaData { 
    id?: string;
    source?: string;
    destination?: string;
    sizeBytes?: number;
    mediaName?: string;
    date?: string;
    season?: number;
    episode?: number;
    mediaType?: MediaFileType;
    deleted?: boolean;
    mediaDesc?: MediaDescriptionData;
}
export namespace MovedMediaData {
}